Module aws_sdk_ssmincidents::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You don't have sufficient access to perform this action.

ConflictException

Updating or deleting a resource causes an inconsistent state.

CreateReplicationSetError

CreateResponsePlanError

CreateTimelineEventError

DeleteIncidentRecordError

DeleteReplicationSetError

DeleteResourcePolicyError

DeleteResponsePlanError

DeleteTimelineEventError

GetIncidentRecordError

GetReplicationSetError

GetResourcePoliciesError

GetResponsePlanError

GetTimelineEventError

InternalServerException

The request processing has failed because of an unknown error, exception or

ListIncidentRecordsError

ListRelatedItemsError

ListReplicationSetsError

ListResponsePlansError

ListTagsForResourceError

ListTimelineEventsError

PutResourcePolicyError

ResourceNotFoundException

Request references a resource which does not exist.

ServiceQuotaExceededException

Request would cause a service quota to be exceeded.

StartIncidentError

TagResourceError

ThrottlingException

The request was denied due to request throttling.

UntagResourceError

UpdateDeletionProtectionError

UpdateIncidentRecordError

UpdateRelatedItemsError

UpdateReplicationSetError

UpdateResponsePlanError

UpdateTimelineEventError

ValidationException

The input fails to satisfy the constraints specified by an AWS service.

Enums

CreateReplicationSetErrorKind

CreateResponsePlanErrorKind

CreateTimelineEventErrorKind

DeleteIncidentRecordErrorKind

DeleteReplicationSetErrorKind

DeleteResourcePolicyErrorKind

DeleteResponsePlanErrorKind

DeleteTimelineEventErrorKind

GetIncidentRecordErrorKind

GetReplicationSetErrorKind

GetResourcePoliciesErrorKind

GetResponsePlanErrorKind

GetTimelineEventErrorKind

ListIncidentRecordsErrorKind

ListRelatedItemsErrorKind

ListReplicationSetsErrorKind

ListResponsePlansErrorKind

ListTagsForResourceErrorKind

ListTimelineEventsErrorKind

PutResourcePolicyErrorKind

StartIncidentErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDeletionProtectionErrorKind

UpdateIncidentRecordErrorKind

UpdateRelatedItemsErrorKind

UpdateReplicationSetErrorKind

UpdateResponsePlanErrorKind

UpdateTimelineEventErrorKind