Module aws_sdk_ssmincidents::model

Modules

add_region_action

See AddRegionAction

delete_region_action

See DeleteRegionAction

empty_chat_channel

See EmptyChatChannel

event_summary

See EventSummary

filter

See Filter

incident_record

See IncidentRecord

incident_record_source

See IncidentRecordSource

incident_record_summary

See IncidentRecordSummary

incident_template

See IncidentTemplate

item_identifier

See ItemIdentifier

region_info

See RegionInfo

region_map_input_value

See RegionMapInputValue

related_item

See RelatedItem

replication_set

See ReplicationSet

resource_policy

See ResourcePolicy

response_plan_summary

See ResponsePlanSummary

ssm_automation

See SsmAutomation

timeline_event

See TimelineEvent

trigger_details

See TriggerDetails

Structs

AddRegionAction

Defines the Region and KMS key to add to the replication set.

DeleteRegionAction

Defines the information about the Region you're deleting from your replication

EmptyChatChannel

Used to remove the chat channel from an incident record or response plan.

EventSummary

Details about a timeline event during an incident.

Filter

Filter the selection by using a condition.

IncidentRecord

The record of the incident that's created when an incident occurs.

IncidentRecordSource

Details about how the incident record was created and when.

IncidentRecordSummary

Details describing an incident record.

IncidentTemplate

Basic details used in creating a response plan. The response plan is then used to create an incident record.

ItemIdentifier

Details and type of a related item.

RegionInfo

Information about a Region in your replication set.

RegionMapInputValue

The mapping between a Region and the key that's used to encrypt the data.

RelatedItem

Resources that responders use to triage and mitigate the incident.

ReplicationSet

The set of Regions that your Incident Manager data will be replicated to and the KMS key used

ResourcePolicy

The resource policy that allows Incident Manager to perform actions on resources on your behalf.

ResponsePlanSummary

Details of the response plan that are used when creating an incident.

SsmAutomation

Details about the Systems Manager automation document that will be used as a runbook during an incident.

TimelineEvent

A significant event that happened during the incident.

TriggerDetails

Details about what caused the incident to be created in Incident Manager.

Enums

Action

AttributeValueList

AutomationExecution

ChatChannel

Condition

IncidentRecordStatus

ItemType

ItemValue

NotificationTargetItem

RegionStatus

RelatedItemsUpdate

ReplicationSetStatus

ResourceType

ServiceCode

SortOrder

SsmTargetAccount

TimelineEventSort

UpdateReplicationSetAction