Module aws_sdk_ssmincidents::operation

Structs

CreateReplicationSet

A replication set replicates and encrypts your data to the provided Regions with the

CreateResponsePlan

Creates a response plan that automates the initial response to incidents. A response

CreateTimelineEvent

Creates a custom timeline event on the incident details page of an incident record.

DeleteIncidentRecord

Delete an incident record from Incident Manager.

DeleteReplicationSet

Deletes all Regions in your replication set. Deleting the replication set deletes all

DeleteResourcePolicy

Deletes the resource policy that

DeleteResponsePlan

Deletes the specified response plan. Deleting a response plan stops all linked

DeleteTimelineEvent

Deletes a timeline event from an incident.

GetIncidentRecord

Returns the details of the specified incident record.

GetReplicationSet

Retrieve your Incident Manager replication set.

GetResourcePolicies

Retrieves the resource policies attached to the specified response plan.

GetResponsePlan

Retrieves the details of the specified response plan.

GetTimelineEvent

Retrieves a timeline event based on its ID and incident record.

ListIncidentRecords

Lists all incident records in your account. Use this command to retrieve the Amazon

ListRelatedItems

List all related items for an incident record.

ListReplicationSets

Lists details about the replication set configured in your account.

ListResponsePlans

Lists all response plans in your account.

ListTagsForResource

Lists the tags that are attached to the specified response plan.

ListTimelineEvents

Lists timeline events of the specified incident record.

PutResourcePolicy

Adds a resource policy to the specified response plan.

StartIncident

Used to start an incident from CloudWatch alarms, EventBridge events, or manually.

TagResource

Adds a tag to a response plan.

UntagResource

Removes a tag from a resource.

UpdateDeletionProtection

Update deletion protection to either allow or deny deletion of the final Region in a

UpdateIncidentRecord

Update the details of an incident record. You can use this action to update an

UpdateRelatedItems

Add or remove related items from the related items tab of an incident record.

UpdateReplicationSet

Add or delete Regions from your replication set.

UpdateResponsePlan

Updates the specified response plan.

UpdateTimelineEvent

Updates a timeline event. You can update events of type Custom